raftstore: extract SnapKey
