+++ components/raftstore/src/store/snap.rs
@@ -1,4 +1,9 @@
+use kvproto::raft_serverpb::RaftSnapshotData;
+use protobuf::Message;
 use std::{error, result};
+use std::fmt::{self, Display, Formatter};
+use std::io::{self, ErrorKind};
+use raft::eraftpb::Snapshot as RaftSnapshot;
 
 quick_error! {
     #[derive(Debug)]
@@ -20,3 +25,45 @@ quick_error! {
 }
 
 pub type Result<T> = result::Result<T, Error>;
+
+#[derive(Clone, Hash, PartialEq, Eq, PartialOrd, Ord, Debug)]
+pub struct SnapKey {
+    pub region_id: u64,
+    pub term: u64,
+    pub idx: u64,
+}
+
+impl SnapKey {
+    #[inline]
+    pub fn new(region_id: u64, term: u64, idx: u64) -> SnapKey {
+        SnapKey {
+            region_id,
+            term,
+            idx,
+        }
+    }
+
+    pub fn from_region_snap(region_id: u64, snap: &RaftSnapshot) -> SnapKey {
+        let index = snap.get_metadata().get_index();
+        let term = snap.get_metadata().get_term();
+        SnapKey::new(region_id, term, index)
+    }
+
+    pub fn from_snap(snap: &RaftSnapshot) -> io::Result<SnapKey> {
+        let mut snap_data = RaftSnapshotData::new();
+        if let Err(e) = snap_data.merge_from_bytes(snap.get_data()) {
+            return Err(io::Error::new(ErrorKind::Other, e));
+        }
+
+        Ok(SnapKey::from_region_snap(
+            snap_data.get_region().get_id(),
+            snap,
+        ))
+    }
+}
+
+impl Display for SnapKey {
+    fn fmt(&self, f: &mut Formatter<'_>) -> fmt::Result {
+        write!(f, "{}_{}_{}", self.region_id, self.term, self.idx)
+    }
+}
