+++ src/raftstore/store/msg.rs
@@ -13,7 +13,7 @@ use raft::SnapshotStatus;
 
 use crate::raftstore::store::fsm::apply::TaskRes as ApplyTaskRes;
 use tikv_misc::store_util::KeysInfoFormatter;
-use crate::raftstore::store::SnapKey;
+use raftstore2::store::snap::SnapKey;
 use tikv_misc::compact_listener::CompactedEvent;
 use tikv_util::escape;
